refactor: styled-component css helper 사용하도록 수정 및 Theme 적용
