+++ src/App.js
@@ -1,12 +1,13 @@
 import Routes from 'Routes';
 import GlobalStyle from 'style/globalStyle';
+import StyleTheme from 'style/theme';
 
 function App() {
   return (
-    <>
+    <StyleTheme>
       <GlobalStyle />
       <Routes />
-    </>
+    </StyleTheme>
   );
 }
